Move AppointmentController functionality into SupplierAsessmentController. This functionality is currently only used for supplier assessment appointments.
